requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.10
Django==5.1.3
django-redis==5.4.0
django-seed==0.3.1
django-silk==5.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==33.1.0
frozenlist==1.5.0
GDAL==3.8.0
gprof2dot==2024.6.6
idna==3.10
multidict==6.1.0
phonenumbers==8.13.50
propcache==0.2.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
toposort==1.10
twilio==9.3.7
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
drf-spectacular~=0.28.0